Expose getBounds seperate to getCenter
