+++ lib/parse.js
@@ -8,6 +8,7 @@ var reflectX = require('./reflect-x')
 var reflectY = require('./reflect-y')
 var toAbsolute = require('./to-absolute')
 var getCenter = require('./get-center')
+var getBounds = require('./get-bounds')
 
 module.exports = function (string, options) {
 
@@ -20,6 +21,7 @@ module.exports = function (string, options) {
   ast.reflectY = reflectY
   ast.toAbsolute = toAbsolute
   ast.getCenter = getCenter
+  ast.getBounds = getBounds
   ast.commands = []
 
   function parseRawString (str) {
@@ -84,4 +86,3 @@ module.exports = function (string, options) {
   return ast
 
 }
-
